Move .dev{n} version suffix handling from setup.py -> GitHub Action
A .dev{n} version suffix is used to distinguish distributions uploaded to
test.pypi.org. Prior to this change the suffix was added in setup.py, which
worked for wheels but failed with when installing an sdist, e.g.

> Package metadata version `33.0.0` does not match given version
> `33.0.0.dev365`

Adding the suffix with egg_info --tag-build avoids this error.

Caveats
1. Prior to this change the dev{n} suffix was added by default.
   After this change no suffix is added by default.
2. Running setup.py directly is deprecated by the setuptools project.

See
- https://setuptools.pypa.io/en/latest/userguide/distribution.html
- https://setuptools.pypa.io/en/latest/deprecated/commands.html#egg-info